Adding base class for viewport windows (preparing for quads)
